Добавлены доп.модули, чтобы весь функционал работал
